+++ common/scala/src/main/resources/application.conf
@@ -143,7 +143,7 @@ whisk {
     }
 
     user-events {
-        enabled = true
+        enabled = false
     }
 
     activation {
