+++ tests/src/test/resources/application.conf.j2
@@ -58,4 +58,7 @@ whisk {
         client-auth = "{{ controller.ssl.clientAuth }}"
       }
     }
+    user-events {
+        enabled = {{ user_events }}
+    }
 }
